[#175271334] refactoring bottomsheet logic in order to condensate the use
