+++ ts/features/bonus/bpd/components/paymentMethodActivationToggle/base/PaymentMethodBpdToggle.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable functional/immutable-data */
-import { useBottomSheetModal } from "@gorhom/bottom-sheet";
 import * as pot from "italia-ts-commons/lib/pot";
 import { Millisecond } from "italia-ts-commons/lib/units";
 import { View } from "native-base";
@@ -9,9 +8,7 @@ import { ImageSourcePropType, StyleSheet } from "react-native";
 import { NavigationContext } from "react-navigation";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import I18n from "../../../../../../i18n";
 import { GlobalState } from "../../../../../../store/reducers/types";
-import { bottomSheetRawConfig } from "../../../../../../utils/bottomSheet";
 import {
   bpdPaymentMethodActivation,
   BpdPaymentMethodActivation,
@@ -20,7 +17,7 @@ import {
   HPan
 } from "../../../store/actions/paymentMethods";
 import { bpdPaymentMethodValueSelector } from "../../../store/reducers/details/paymentMethods";
-import { BpdChangeActivationConfirmationScreen } from "../confirm/BpdChangeActivationConfirmationScreen";
+import { useChangeActivationConfirmationBottomSheet } from "../bottomsheet/BpdChangeActivationConfirmationScreen";
 import { BpdToggle } from "./BpdToggle";
 import { PaymentMethodRepresentation } from "./PaymentMethodRepresentation";
 
@@ -153,29 +150,8 @@ const PaymentMethodActivationToggle: React.FunctionComponent<Props> = props => {
     useInitialValue(props);
   }
 
-  const { present, dismiss } = useBottomSheetModal();
-
-  const openModalBox = (newVal: boolean) => {
-    const bottomSheetProps = bottomSheetRawConfig(
-      <BpdChangeActivationConfirmationScreen
-        onCancel={dismiss}
-        onConfirm={() => {
-          dismiss();
-          props.updateValue(props.hPan, newVal);
-        }}
-        type={newVal ? "Activation" : "Deactivation"}
-        representation={{ caption: props.caption, icon: props.icon }}
-      />,
-      newVal
-        ? I18n.t("bonus.bpd.details.paymentMethods.activate.title")
-        : I18n.t("bonus.bpd.details.paymentMethods.deactivate.title"),
-      466,
-      dismiss
-    );
-    present(bottomSheetProps.content, {
-      ...bottomSheetProps.config
-    });
-  };
+  const askConfirmation = useChangeActivationConfirmationBottomSheet(props)
+    .present;
 
   return (
     <>
@@ -186,7 +162,9 @@ const PaymentMethodActivationToggle: React.FunctionComponent<Props> = props => {
         />
         <BpdToggle
           graphicalValue={graphicalState}
-          onValueChanged={openModalBox}
+          onValueChanged={newVal =>
+            askConfirmation(newVal, () => props.updateValue(props.hPan, newVal))
+          }
           // TODO: when onPress -> show bottomsheet explaining why the bpd cannot be activated
         />
       </View>
